+++ src/main/ipc/engines.ts
@@ -12,7 +12,11 @@ import {
   handleUpdateEngineAlias,
   loadSavedEngines
 } from './engineHandlers'
-import { clearEnginePluginCache } from './enginePlugins'
+import {
+  clearEnginePluginCache,
+  getEnginePluginCacheTTL,
+  setEnginePluginCacheTTL
+} from './enginePlugins'
 import type { LaunchConfig } from '../utils/launchConfigArgs'
 
 /**
@@ -66,15 +70,10 @@ export function registerEngineHandlers(ipcMain_: typeof ipcMain): void {
   })
 
   ipcMain_.handle('get-engine-plugin-cache-ttl', (): number => {
-    // lazy import to avoid cycles
-    // eslint-disable-next-line @typescript-eslint/no-require-imports
-    const { getEnginePluginCacheTTL } = require('./enginePlugins')
     return getEnginePluginCacheTTL()
   })
 
   ipcMain_.handle('set-engine-plugin-cache-ttl', (_event, ms: number) => {
-    // eslint-disable-next-line @typescript-eslint/no-require-imports
-    const { setEnginePluginCacheTTL } = require('./enginePlugins')
     setEnginePluginCacheTTL(Number(ms) || 0)
   })